拆分主窗口布局与窗口操作
